modific and added some changes in module vehicle and users

+++ internal/app/application/vehicles/deleteVehicle.go
@@ -3,11 +3,13 @@ package vehicles
 import (
 	"context"
 	"errors"
+	"fmt"
 	"shipping-app/internal/app/domain/ports/repository"
 )
 
 var (
-	ErrVehicleNotFoundDelete = errors.New("Vehículo no encontrado")
+	ErrVehicleNotFoundDelete  = errors.New("Vehículo no encontrado")
+	ErrVehicleHasActiveOrders = errors.New("vehicle has active orders and cannot be deleted")
 )
 
 type DeleteVehicleUseCase struct {
@@ -28,5 +30,13 @@ func (uc *DeleteVehicleUseCase) Execute(ctx context.Context, id uint) error {
 		return ErrVehicleNotFound
 	}
 
+	hasActiveOrders, err := uc.repo.HasActiveVehicleInOrder(ctx, id)
+	if err != nil {
+		return fmt.Errorf("error checking active orders: %w", err)
+	}
+
+	if hasActiveOrders {
+		return ErrVehicleHasActiveOrders
+	}
 	return uc.repo.DeleteVehicle(id)
 }
